pyproject.toml
[project]
name = "price-platform"
version = "0.1.0"
description = "Shared webapp and auth foundation for product price applications"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=23.1.0",
    "beautifulsoup4>=4.12.0",
    "docopt-ng>=0.9.0",
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    "my-lib @ git+https://github.com/kimata/my-py-lib@22be9e19b8d3ee804d492bb1db2e2938f08cbb8a",
    "PyJWT>=2.8.0",
    "requests>=2.32.0",
    "selenium>=4.0.0",
]
authors = [
    { name = "KIMATA Tetsuya", email = "kimata@green-rabbit.net" }
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pyright>=1.1.408",
    "pytest>=9.0.0",
    "ty>=0.0.1a7",
    "types-pyyaml>=6.0.12",
    "types-requests>=2.32.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/price_platform"]
include = ["src/price_platform/schema/*.schema"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
default-groups = ["dev"]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "^tests/",
    "^.venv/",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
include = ["src", "tests"]
reportMissingImports = false

[tool.ty]
[tool.ty.rules]
unresolved-import = "ignore"

[[tool.ty.overrides]]
include = ["src/price_platform/store/_price_event_suppression.py"]
[tool.ty.overrides.rules]
invalid-argument-type = "warn"

[[tool.ty.overrides]]
include = ["src/price_platform/store/price_event_detector.py"]
[tool.ty.overrides.rules]
invalid-assignment = "warn"

[[tool.ty.overrides]]
include = ["src/price_platform/webapp/app_factory.py"]
[tool.ty.overrides.rules]
invalid-assignment = "warn"

[tool.pytest.ini_options]
testpaths = ["tests"]
